add services for app
