+++ src/main/java/home/task/spring/controller/CustomerController.java
@@ -1,7 +1,7 @@
 package home.task.spring.controller;
 
 import home.task.spring.model.Customer;
-import home.task.spring.services.TaskService;
+import home.task.spring.services.CustomerService;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -14,12 +14,11 @@ import java.util.List;
 @RequestMapping("/")
 public class CustomerController {
     @Autowired
-    TaskService taskService;
-
+    private CustomerService customerService;
 
     @GetMapping("customers")
     public List<Customer> getAllCustomer() {
-        return taskService.getAllCustomer();
+        return customerService.getAllCustomer();
     }
 
     @ApiOperation(value = "Find customer by id.",
@@ -27,7 +26,7 @@ public class CustomerController {
             response = Customer.class)
     @GetMapping("customer/{id}")
     public Customer getCustomerById(@PathVariable Long id) {
-        return taskService.getCustomerById(id);
+        return customerService.getCustomerById(id);
     }
 
     @ApiOperation(value = "Add new customer",
@@ -35,7 +34,7 @@ public class CustomerController {
             response = HttpStatus.class)
     @PostMapping("customer")
     public void addCustomer(@RequestBody Customer customer) {
-        taskService.addCustomer(customer);
+        customerService.addCustomer(customer);
     }
 
     @ApiOperation(value = "Delete  customer",
@@ -43,6 +42,6 @@ public class CustomerController {
             response = HttpStatus.class)
     @DeleteMapping("customer/{id}")
     public void deleteCustomer(@PathVariable Long id) {
-        taskService.deleteCustomer(id);
+        customerService.deleteCustomer(id);
     }
 }
