+++ src/main/java/home/task/spring/services/impl/CustomerServiceImpl.java
@@ -1,0 +1,37 @@
+package home.task.spring.services.impl;
+
+import home.task.spring.model.Customer;
+import home.task.spring.repository.CustomerRepository;
+import home.task.spring.services.CustomerService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+@Transactional
+public class CustomerServiceImpl implements CustomerService {
+    @Autowired
+    CustomerRepository customerRepository;
+
+    @Override
+    public List<Customer> getAllCustomer() {
+        return customerRepository.getAllCustomer();
+    }
+
+    @Override
+    public Customer getCustomerById(Long id) {
+        return customerRepository.getCustomerById(id);
+    }
+
+    @Override
+    public void addCustomer(Customer customer) {
+        customerRepository.addCustomer(customer);
+    }
+
+    @Override
+    public void deleteCustomer(Long id) {
+        customerRepository.deleteCustomer(id);
+    }
+}
