--- src/main/webapp/WEB-INF/web.xml
@@ -1,8 +1,0 @@
-<html>
-    <head>
-        <title>Title</title>
-    </head>
-    <body>
-
-    </body>
-</html>
